+++ internal/delivery/http/v1/order.go
@@ -36,6 +36,12 @@ func (h *Handler) initOrderRoutes(api *echo.Group) {
 		users.Use(h.identity)
 		users.GET("/:id/orders", h.usersGetAllOrders)
 	}
+
+	couriers := api.Group("/couriers")
+	{
+		couriers.Use(h.identity)
+		couriers.GET("/:cid/orders", h.getActiveCourierOrder)
+	}
 }
 
 type orderInput struct {
@@ -400,4 +406,36 @@ func (h *Handler) usersGetAllOrders(ctx echo.Context) error {
 	}
 
 	return ctx.JSON(http.StatusOK, orders)
+}
+
+// @Summary Get Active Order
+// @Security CourierAuth
+// @Tags couriers
+// @Description get active order for courier
+// @ModuleID getActiveCourierOrder
+// @Accept  json
+// @Produce  json
+// @Param cid path string true "Courier id"
+// @Success 200 {object} response
+// @Failure 400,403,404 {object} response
+// @Failure 500 {object} response
+// @Failure default {object} response
+// @Router /couriers/{cid}/orders [get]
+func (h *Handler) getActiveCourierOrder(ctx echo.Context) error {
+	clientId, clientType, err := h.getClientParams(ctx)
+	if err != nil {
+		return newResponse(ctx, http.StatusInternalServerError, err.Error())
+	}
+
+	courierId, err := strconv.Atoi(ctx.Param("cid"))
+	if err != nil || courierId == 0 {
+		return newResponse(ctx, http.StatusBadRequest, "Invalid restaurantId")
+	}
+
+	order, err := h.services.Order.GetActiveOrder(clientId, clientType, courierId)
+	if err != nil {
+		return newResponse(ctx, http.StatusInternalServerError, err.Error())
+	}
+
+	return ctx.JSON(http.StatusOK, order)
 }
