+++ internal/repository/repository.go
@@ -20,7 +20,6 @@ type Courier interface {
 	GetByCredentials(phone, password string) (*domain.Courier, error)
 	GetById(courierId int) (*domain.Courier, error)
 	Update(courierId int, input *domain.Courier) error
-	GetActiveOrder(courierId int) (*domain.Order, error)
 }
 
 type Restaurant interface {
@@ -48,6 +47,7 @@ type Order interface {
 	GetItemById(orderItemId int) (*domain.OrderItem, error)
 	UpdateItem(orderItemId, menuItemsCount int) error
 	DeleteItem(orderItemId int, orderId int) error
+	GetActiveOrder(courierId int) (*domain.Order, error)
 }
 
 type MenuItem interface {
